Consumers tracking cleanup
